// src/dfs/generated/handlers/IFileSystemOperationsHandler.ts
/*
 * Copyright (c) Microsoft Corporation. All rights reserved.
 * Licensed under the MIT License. See License.txt in the project root for
 * license information.
 *
 * Code generated by Microsoft (R) AutoRest Code Generator.
 * Changes may cause incorrect behavior and will be lost if the code is
 * regenerated.
 */
// tslint:disable:max-line-length

import * as Models from "../artifacts/models";
import Context from "../Context";

export default interface IFileSystemOperationsHandler {
  create(options: Models.FileSystemCreateOptionalParams, context: Context): Promise<Models.FileSystemCreateResponse>;
  setProperties(options: Models.FileSystemSetPropertiesOptionalParams, context: Context): Promise<Models.FileSystemSetPropertiesResponse>;
  getProperties(options: Models.FileSystemGetPropertiesOptionalParams, context: Context): Promise<Models.FileSystemGetPropertiesResponse>;
  delete(options: Models.FileSystemDeleteMethodOptionalParams, context: Context): Promise<Models.FileSystemDeleteResponse>;
  listPaths(recursive: boolean, options: Models.FileSystemListPathsOptionalParams, context: Context): Promise<Models.FileSystemListPathsResponse>;
  listBlobFlatSegment(options: Models.FileSystemListBlobFlatSegmentOptionalParams, context: Context): Promise<Models.FileSystemListBlobFlatSegmentResponse>;
  listBlobHierarchySegment(delimiter: string, options: Models.FileSystemListBlobHierarchySegmentOptionalParams, context: Context): Promise<Models.FileSystemListBlobHierarchySegmentResponse>;
}